Refactor mongodb connection to own pagkage
